Improved loopback test suite to support multiple configuration.
To run the test suite:

./run_test.sh sample_format sample_rate channels duration
Where:
    sample_format can be one of S16_LE, S24_3LE, S32_LE
    sample_rate can be one of 44100, 48000, 96000
    channels can be one of 1, 2, 4
    duration is in the range 1 to 10 minutes

The test suite creates a raw file with the specified parameters, runs a loopback test where the file gets played and recorded using the loopback network interface and checks that the recorded file, after the initial silence, contains the expected samples sequence.
This test was developed to further investigate the issue #17.