Fix HttpServer::init() erroneous return value
Fix bug HttpServer::init() would incorrectly return true when the
server is not responding, and sometimes would incorrectly return false
when the server is responding.

Using postfix decrement in the while condition caused the returned value
to be one less than intended, making an intended 0 value actually -1
which was cast to true, and an intended 1 value actually 0 and returned
as false.