Compile the daemon in parallel.
The compilation can be a bit slow when done sequentially. The `-j`
option automatically picks the number of parallel build jobs based on
the number of available cores.

The parallelization might be a concern then building on systems with a
small amount of RAM, that can be addressed when such occasion arises.

As a side note, in many build systems, we leave the `make` stage to be
executed separately by the user, so that the user can pick the level of
parallelization. There are a few invocations of `make` in the build.sh
script, I'm adding the `-j` flag only to the one that was slow in
practice.