aes67-daemon: Make dependency on libfaac optional
Previously, since the introduction of the streamer support in version 2.0.0,
it was impossible to compile the daemon without a dependency on libfaac.
This prevented the daemon from being built on Debian systems without `non-free`
repositories enabled. It seems the issue with libfaac is that it is based on
code with a non-DFSG compliant license.

This commit introduces a new config option WITH_STREAMER which is ON by default,
causing the behaviour to stay backwards-compatible with the one before this
commit. It is newly possible to pass `-DWITH_STREAMER=OFF` to cmake to build
without streamer support and without the dependency on libfaac.